FIX: return a error if a product doesn't exist

+++ src/controllers/productController.ts
@@ -14,6 +14,9 @@ export const productController = {
     getById: async (req: Request, res: Response) => {
         try {
             const product = await productService.getById(parseInt(req.params.id));
+            if (!product) {
+                return res.status(404).json({ error: 'Product not found' });
+            }
             return res.status(201).json(product);
         } catch (e) {
             res.status(500).json({ error: (e as Error).message });
@@ -29,6 +32,10 @@ export const productController = {
     },
     update: async (req: Request, res: Response) => {
         try {
+            const productExist = await productService.getById(parseInt(req.params.id));
+            if (!productExist) {
+                return res.status(404).json({ error: 'Product not found' });
+            }
             const product = await productService.update(parseInt(req.params.id), req.body);
             return res.status(200).json(product);
         } catch (e) {
@@ -37,6 +44,10 @@ export const productController = {
     },
     delete: async (req: Request, res: Response) => {
         try {
+            const productExist = await productService.getById(parseInt(req.params.id));
+            if (!productExist) {
+                return res.status(404).json({ error: 'Product not found' });
+            }
             const product = await productService.delete(parseInt(req.params.id));
             return res.status(200).json(product);
         } catch (e) {
